ref: split large assert files into separate ones
