+++ tests/api/blocks_test.rs
@@ -1,10 +1,9 @@
+use crate::utils::asserts::block::{assert_block_data, test_block_array};
+use crate::utils::asserts::meta::assert_meta;
+use crate::utils::asserts::transaction::test_transaction_array;
+use crate::utils::mockito_helpers::{mock_client, mock_http_request};
 use serde_json::{from_str, Value};
 use std::borrow::Borrow;
-
-use crate::utils::assert_helpers::{
-    assert_block_data, assert_meta, test_block_array, test_transaction_array,
-};
-use crate::utils::mockito_helpers::{mock_client, mock_http_request};
 
 #[test]
 fn test_blocks_all() {
