+++ tests/api/peers_test.rs
@@ -1,12 +1,10 @@
-use serde_json::{from_str, Value};
-
+use crate::utils::asserts::meta::assert_meta;
+use crate::utils::asserts::peer::{assert_peer_data, test_peer_array};
+use crate::utils::mockito_helpers::{mock_client, mock_http_request};
 use arkecosystem_client::api::models::peer::Peer;
-
 use arkecosystem_client::api::models::shared::Response;
 use arkecosystem_client::Connection;
-
-use crate::utils::assert_helpers::{assert_meta, assert_peer_data, test_peer_array};
-use crate::utils::mockito_helpers::{mock_client, mock_http_request};
+use serde_json::{from_str, Value};
 use std::borrow::Borrow;
 
 #[test]
