+++ tests/api/wallets_tests.rs
@@ -1,10 +1,9 @@
+use crate::utils::asserts::meta::assert_meta;
+use crate::utils::asserts::transaction::{test_lock_array, test_transaction_array};
+use crate::utils::asserts::wallet::{assert_wallet_data, test_wallet_array};
+use crate::utils::mockito_helpers::{mock_client, mock_http_request, mock_post_request};
 use serde_json::{from_str, Value};
 use std::borrow::Borrow;
-
-use crate::utils::assert_helpers::{
-    assert_meta, assert_wallet_data, test_lock_array, test_transaction_array, test_wallet_array,
-};
-use crate::utils::mockito_helpers::{mock_client, mock_http_request, mock_post_request};
 use std::collections::HashMap;
 
 #[test]
