+++ tests/utils/asserts/mod.rs
@@ -1,0 +1,10 @@
+pub mod block;
+pub mod delegate;
+pub mod meta;
+pub mod node;
+pub mod peer;
+pub mod shared;
+pub mod transaction;
+pub mod transaction_fees;
+pub mod transaction_types;
+pub mod wallet;
